+++ resolwe_bio/tests/processes/test_support_processors.py
@@ -557,6 +557,60 @@ re-save-file report "${NAME}".txt
         )
         self.assertFileExists(multiqc, "report")
 
+    @tag_process("multiqc")
+    def test_multiqc_chipqc(self):
+        with self.preparation_stage():
+
+            def set_sample_name(data, sample_name):
+                """Set sample name."""
+                sample = Sample.objects.get(data=data)
+                sample.name = sample_name
+                sample.save()
+
+            process = Process.objects.create(
+                name="Upload chipqc data mock process",
+                requirements={
+                    "expression-engine": "jinja",
+                    "resources": {"network": True,},
+                    "executor": {"docker": {"image": "resolwebio/base:ubuntu-18.04",},},
+                },
+                entity_type="sample",
+                entity_descriptor_schema="sample",
+                contributor=self.contributor,
+                type="data:chipqc:",
+                input_schema=[{"name": "src", "type": "basic:file:",},],
+                output_schema=[
+                    {"name": "ccplot", "type": "basic:file:",},
+                    {"name": "coverage_histogram", "type": "basic:file:",},
+                    {"name": "peak_profile", "type": "basic:file:",},
+                    {"name": "peaks_barplot", "type": "basic:file:",},
+                    {"name": "peaks_density_plot", "type": "basic:file:",},
+                ],
+                run={
+                    "language": "bash",
+                    "program": r"""
+re-import {{ src.file_temp|default(src.file) }} {{ src.file }} "png" "png" 0.1 extract
+re-save-file ccplot "${NAME}".png
+re-save-file coverage_histogram "${NAME}".png
+re-save-file peak_profile "${NAME}".png
+re-save-file peaks_barplot "${NAME}".png
+re-save-file peaks_density_plot "${NAME}".png
+""",
+                },
+            )
+
+            chipqc = self.run_process(
+                process.slug,
+                {"src": os.path.join("chipqc", "output", "PeakProfile_mqc.png"),},
+            )
+            set_sample_name(chipqc, "ChipQC test")
+
+        multiqc = self.run_process(
+            "multiqc",
+            {"data": [chipqc.id,], "advanced": {"dirs": True, "config": True,}},
+        )
+        self.assertFileExists(multiqc, "report")
+
     @tag_process("seqtk-sample-single", "seqtk-sample-paired")
     def test_seqtk_sample(self):
         with self.preparation_stage():